직원상세정보, 직원정보 컴포넌트 분리_이슬기_240306
