+++ src/pages/admin/EmpDetail.jsx
@@ -4,9 +4,9 @@ import styles from './empDetailInfo.module.css';
 import { getEmpList, saveEmpDetails, setDetail } from '../../redux/empInfosSlice';
 import { Col, Row } from 'antd';
 import { DeptNameDB } from '../../services/api/empCreateApi';
-import styled from 'styled-components';
 import EmpUploadImg from './EmpUploadImg';
 import { JobListDB } from '../../services/api/deptApi';
+import EmpDetailInputField from './EmpDetailInputField';
 
 const EmpDetail = () => {
   const dispatch = useDispatch();
@@ -41,8 +41,7 @@ const EmpDetail = () => {
     }
   }, [deptCd]);
 
-    const deptName = () => {
-    console.log("deptName");
+  const deptName = () => {
     DeptNameDB()
       .then((response) => {
         setDept(response);
@@ -52,14 +51,11 @@ const EmpDetail = () => {
       });
   };
 
-  // deptName에 있는 CD와 같은 CD 가져오기
   const deptJob = () => {
-    console.log("deptJob");
     const data = {
       CD : deptCd,
       MOD_ID: empData.e_no
     }
-    console.log(data);
     JobListDB(data)
       .then((response) => {
         setJob(response.data);
@@ -68,22 +64,6 @@ const EmpDetail = () => {
         console.log(error);
       })
   } 
-
-  const passwordGenerate = () => {
-    const chars =
-      "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%&*";
-    let randomStr = "";
-    for (let i = 0; i < 10; i++) {
-      let randomIndex = Math.floor(Math.random() * chars.length);
-      randomStr += chars[randomIndex];
-    }
-    console.log(randomStr);
-    setPassword(randomStr);
-    setUpdatedEmployee(prevState => ({
-      ...prevState,
-      E_PASSWORD: randomStr // E_PASSWORD 항목만 업데이트
-    }))
-  };
 
   const handleEdit = () => {
     setEditing(true); // 수정 버튼을 누를 때 수정 모드 활성화
@@ -109,6 +89,15 @@ const EmpDetail = () => {
     }));
   }
 
+  // 비밀번호 재발급 처리
+  const handlePasswordGenerate = (newPassword) => {
+    setPassword(newPassword);
+    setUpdatedEmployee(prevState => ({
+      ...prevState,
+      E_PASSWORD: newPassword // E_PASSWORD 항목만 업데이트
+    }))
+  };
+
   // 수정된 직원 정보 저장 후, 전체 직원 목록을 다시 가져옴
   const handleSaveChanges = () => {
     dispatch(saveEmpDetails(updatedEmployee)) // 수정된 직원 정보 저장
@@ -116,7 +105,6 @@ const EmpDetail = () => {
         dispatch(setDetail(updatedEmployee)); // 리덕스 스토어에서 선택된 직원 정보 업데이트
         setEditing(false); // 저장 후 수정 모드 비활성화
         dispatch(getEmpList()); // 저장 후 전체 목록 갱신
-
         // 수정된 직원 정보를 UI에 반영하기 위해 상태 업데이트
         setOriginalEmployee(updatedEmployee);
       })
@@ -124,74 +112,6 @@ const EmpDetail = () => {
         console.error('Error saving employee details: ', error);
       });
   }
-
-  const renderInputField = ({ label, name, type, options }, index) => {
-    if (name === 'E_PASSWORD') {
-      return (
-        <div className={styles.empInfoItem} key={name}>
-          <div className={styles.label}>{label}</div>
-          <div className={styles.selectContainer}>
-            <input
-              className={styles.inputFields}
-              type="password"
-              value={updatedEmployee[name] || ''} // updatedEmployee의 비밀번호 값으로 설정
-              onChange={handleInputChange} // 입력 필드가 변경되면 상태를 업데이트
-              readOnly={!editing}
-              name={name}
-            />
-            <MyButton type="button" onClick={passwordGenerate}>
-              임시비밀번호재발급
-            </MyButton>
-          </div>
-          {index !== inputFields.length - 1 && <div className={styles.divider} />}
-        </div>
-      );
-    } else {
-      return (
-        <div className={styles.empInfoItem} key={name}>
-          <div className={styles.label}>{label}</div>
-          <div className={styles.selectContainer}>
-            {type === 'select' ? (
-              <select
-                className={styles.selectBox}
-                value={updatedEmployee[name] || ''}
-                onChange={handleInputChange}
-                disabled={!editing}
-                name={name}
-              >
-                {name === 'DEPT_NAME' ? (
-                  // 부서 선택 옵션을 동적으로 가져오기
-                  dept.map((item, index) => (
-                    <option key={index} value={item.CD_VALUE}>{item.CD_VALUE}</option>
-                  ))
-                ) : name === 'E_OCCUP' ? (
-                  // 직종 선택 옵션을 동적으로 가져오기
-                  job.map((item, index) => (
-                    <option key={index} value={item.CD_VALUE}>{item.CD_VALUE}</option>
-                  ))
-                ) : (
-                  // 기존의 옵션들은 그대로 사용
-                  options.map((option, index) => (
-                    <option key={index} value={option}>{option}</option>
-                  ))
-                )}
-              </select>
-            ) : (
-              <input
-                className={styles.inputFields}
-                type={type}
-                value={updatedEmployee[name] || ''}
-                onChange={handleInputChange}
-                readOnly={!editing}
-                name={name}
-              />
-            )}
-          </div>
-          {index !== inputFields.length - 1 && <div className={styles.divider} />}
-        </div>
-      );
-    }
-  };
   
   const inputFields = [
     { label: '사원명', name: 'E_NAME', type: 'text' },
@@ -204,7 +124,7 @@ const EmpDetail = () => {
     { label: '이메일', name: 'E_EMAIL', type: 'text' },
     { label: '주소', name: 'E_ADDRESS', type: 'text' },
     { label: '부서', name: 'DEPT_NAME', type: 'select', options: dept.map(item => item.CD_VALUE) },
-    { label: '비밀번호', name: 'E_PASSWORD', type: 'password' },
+    { label: '비밀번호', name: 'E_PASSWORD', type: 'text' },
     { label: '권한', name: 'E_AUTH', type: 'select', options: ['ADMIN', 'USERA', 'USERB'] },
     { label: '현황', name: 'E_STATUS', type: 'select', options: ['재직', '휴직', '퇴직'] },
     { label: '직종', name: 'E_OCCUP', type: 'select', options: job.map(item => item.CD_VALUE) },
@@ -227,7 +147,18 @@ const EmpDetail = () => {
             { editing && <EmpUploadImg imageUrlChange={handleImageUrlChange}/> }
           </div>
         </div>
-        {inputFields.map(renderInputField)}
+        {inputFields.map((field, index) => (
+          <EmpDetailInputField
+            key={index}
+            field={field}
+            updatedEmployee={updatedEmployee}
+            handleInputChange={handleInputChange}
+            editing={editing}
+            dept={dept}
+            job={job}
+            onPasswordGenerate={handlePasswordGenerate} // 비밀번호 생성 함수 전달
+          />
+        ))}
       </div>
     <Row style={{ marginBottom: "10px" }}>
       <Col md={15}></Col>
@@ -251,10 +182,3 @@ const EmpDetail = () => {
 };
 
 export default EmpDetail;
-
-const MyButton = styled.button`
-  margin-top: 35px;
-  border-radius: 5px;
-  background-color: grey;
-  color: white;
-`;
